+++ test/images/agnhost/BUILD
@@ -13,17 +13,14 @@ go_binary(
 
 go_library(
     name = "go_default_library",
-    srcs = [
-        "agnhost.go",
-        "common.go",
-        "utils.go",
-        "utils_windows.go",
-    ],
+    srcs = ["agnhost.go"],
     importpath = "k8s.io/kubernetes/test/images/agnhost",
     deps = [
+        "//test/images/agnhost/dns:go_default_library",
         "//test/images/agnhost/net:go_default_library",
         "//test/images/agnhost/netexec:go_default_library",
         "//test/images/agnhost/nettest:go_default_library",
+        "//test/images/agnhost/pause:go_default_library",
         "//test/images/agnhost/webhook:go_default_library",
         "//vendor/github.com/spf13/cobra:go_default_library",
         "//vendor/k8s.io/klog:go_default_library",
@@ -41,9 +38,11 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
+        "//test/images/agnhost/dns:all-srcs",
         "//test/images/agnhost/net:all-srcs",
         "//test/images/agnhost/netexec:all-srcs",
         "//test/images/agnhost/nettest:all-srcs",
+        "//test/images/agnhost/pause:all-srcs",
         "//test/images/agnhost/webhook:all-srcs",
     ],
     tags = ["automanaged"],
