+++ test/images/agnhost/agnhost.go
@@ -22,54 +22,23 @@ import (
 	"github.com/spf13/cobra"
 
 	"k8s.io/klog"
+	"k8s.io/kubernetes/test/images/agnhost/dns"
 	"k8s.io/kubernetes/test/images/agnhost/net"
 	"k8s.io/kubernetes/test/images/agnhost/netexec"
 	"k8s.io/kubernetes/test/images/agnhost/nettest"
+	"k8s.io/kubernetes/test/images/agnhost/pause"
 	"k8s.io/kubernetes/test/images/agnhost/webhook"
 )
 
 func main() {
-	cmdDNSSuffix := &cobra.Command{
-		Use:   "dns-suffix",
-		Short: "Prints the host's DNS suffix list",
-		Long:  `Prints the DNS suffixes of this host.`,
-		Args:  cobra.MaximumNArgs(0),
-		Run:   printDNSSuffixList,
-	}
-
-	cmdDNSServerList := &cobra.Command{
-		Use:   "dns-server-list",
-		Short: "Prints the host's DNS Server list",
-		Long:  `Prints the DNS Server list of this host.`,
-		Args:  cobra.MaximumNArgs(0),
-		Run:   printDNSServerList,
-	}
-
-	cmdEtcHosts := &cobra.Command{
-		Use:   "etc-hosts",
-		Short: "Prints the host's /etc/hosts file",
-		Long:  `Prints the "hosts" file of this host."`,
-		Args:  cobra.MaximumNArgs(0),
-		Run:   printHostsFile,
-	}
-
-	cmdPause := &cobra.Command{
-		Use:   "pause",
-		Short: "Pauses",
-		Long:  `Pauses the execution. Useful for keeping the containers running, so other commands can be executed.`,
-		Args:  cobra.MaximumNArgs(0),
-		Run:   pause,
-	}
-
 	rootCmd := &cobra.Command{Use: "app"}
-	rootCmd.AddCommand(cmdDNSSuffix)
-	rootCmd.AddCommand(cmdDNSServerList)
-	rootCmd.AddCommand(cmdEtcHosts)
-	rootCmd.AddCommand(cmdPause)
-
+	rootCmd.AddCommand(dns.CmdDNSSuffix)
+	rootCmd.AddCommand(dns.CmdDNSServerList)
+	rootCmd.AddCommand(dns.CmdEtcHosts)
 	rootCmd.AddCommand(net.CmdNet)
 	rootCmd.AddCommand(netexec.CmdNetexec)
 	rootCmd.AddCommand(nettest.CmdNettest)
+	rootCmd.AddCommand(pause.CmdPause)
 	rootCmd.AddCommand(webhook.CmdWebhook)
 
 	// NOTE(claudiub): Some tests are passing logging related flags, so we need to be able to
