+++ test/images/agnhost/dns/dns.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package dns
 
 import (
 	"strings"
